+++ packages/@storefront/flux-capacitor/src/core/actions/validators.ts
@@ -71,8 +71,15 @@ export const notOnFirstPage: Validator = {
 };
 
 export const isRefinementDeselectedByIndex: Validator<Actions.Payload.Navigation.Refinement> = {
-  func: ({ navigationId, index }, state) => Selectors.isRefinementDeselected(state, navigationId, index),
-  msg: 'navigation does not exist or refinement is already selected'
+  // tslint:disable-next-line max-line-length
+  func: ({ navigationId, index }, state) => !!navigationId && Selectors.isRefinementDeselected(state, navigationId, index),
+  msg: 'navigation does not exist or refinements are already selected'
+};
+
+export const areMultipleRefinementsDeselectedByIndex: Validator<Actions.Payload.Navigation.MultipleRefinements> = {
+  // tslint:disable-next-line max-line-length
+  func: ({ navigationId, indices }, state) =>  !!navigationId && Array.isArray(indices) && indices.some((index) => Selectors.isRefinementDeselected(state, navigationId, index)),
+  msg: 'navigation does not exist or all refinements are already selected'
 };
 
 export const isRefinementSelectedByIndex: Validator<Actions.Payload.Navigation.Refinement> = {
@@ -83,6 +90,13 @@ export const isRefinementSelectedByIndex: Validator<Actions.Payload.Navigation.R
 export const isPastPurchaseRefinementDeselectedByIndex: Validator<Actions.Payload.Navigation.Refinement> = {
   func: ({ navigationId, index }, state) => Selectors.isPastPurchaseRefinementDeselected(state, navigationId, index),
   msg: 'navigation does not exist or refinement is already selected'
+};
+
+  // tslint:disable-next-line max-line-length
+export const areMultiplePastPurchaseRefinementsDeselectedByIndex: Validator<Actions.Payload.Navigation.MultipleRefinements> = {
+  // tslint:disable-next-line max-line-length
+  func: ({ navigationId, indices }, state) => Array.isArray(indices) && indices.some((index) => Selectors.isPastPurchaseRefinementDeselected(state, navigationId, index)),
+  msg: 'navigation does not exist or all refinements are already selected'
 };
 
 export const isPastPurchaseRefinementSelectedByIndex: Validator<Actions.Payload.Navigation.Refinement> = {