+++ packages/@storefront/flux-capacitor/test/unit/core/reducers/data/navigations.ts
@@ -199,13 +199,52 @@ suite('navigations', ({ expect }) => {
       expect(reducer).to.eql(newState);
     });
 
-    it('should return state on SELECT_REFINEMENT if no navigationId and refinementIndex', () => {
-      const reducer = navigations(state, <any>{
-        type: Actions.SELECT_REFINEMENT,
-        payload: {}
+    it('should add selected refinement state on SELECT_MULTIPLE_REFINEMENTS', () => {
+      const newState = {
+        ...state,
+        byId: {
+          Format,
+          Section: {
+            ...Section,
+            selected: [3, 0, 4, 9],
+          },
+          Department,
+        },
+      };
+
+      const reducer = navigations(state, {
+        type: Actions.SELECT_MULTIPLE_REFINEMENTS,
+        payload: {
+          navigationId: 'Section',
+          indices: [0, 4, 9]
+        }
       });
 
-      expect(reducer).to.eql(state);
+      expect(reducer).to.eql(newState);
+    });
+
+    it('should not add duplicate indices on SELECT_MULTIPLE_REFINEMENTS', () => {
+      const newState = {
+        ...state,
+        byId: {
+          Format,
+          Section: {
+            ...Section,
+            selected: [3, 0, 4, 9],
+          },
+          Department,
+        },
+      };
+
+      const reducer = navigations(state, {
+        type: Actions.SELECT_MULTIPLE_REFINEMENTS,
+        payload: {
+          navigationId: 'Section',
+          indices: [0, 4, 9, 3]
+        }
+      });
+
+      expect(reducer).to.eql(newState);
     });
 
     it('should not add duplicate range refinement on ADD_REFINEMENT', () => {
@@ -424,7 +463,7 @@ suite('navigations', ({ expect }) => {
         payload: {}
       });
 
-      expect(reducer).to.eql(state);
+      expect(reducer).to.eq(state);
     });
 
     it('should update refinements state on RECEIVE_MORE_REFINEMENTS', () => {
@@ -465,13 +504,13 @@ suite('navigations', ({ expect }) => {
         payload: {}
       });
 
-      expect(reducer).to.eql(state);
+      expect(reducer).to.eq(state);
     });
 
     it('should return state on default', () => {
       const reducer = navigations(state, <any>{});
 
-      expect(reducer).to.eql(state);
+      expect(reducer).to.eq(state);
     });
   });
 });